Refactor StandardExecutionManager with Popen and Thread
The previous approach based on asyncio was failing in combination
with QThread, because asyncio is meant to be called from the main
thread for some reason (or at least that's the impression I got,
first time user here).

The new approach is simpler and seems to work just fine for us.