Towards multiprocess execution
Re #10

The idea here is that the new SpineEngineExperimental class would
handle multiprocess execution whenever we manage to make it fast.

At the moment it works, pipeline execution is parallelized across
solids whenever possible. The problem is, dagster needs to share
solid outputs across processes, and for that we're using
file system storage (the simplest one I found).
As result, there's a lot of
file writing and reading going on which makes it even slower than
single process execution. For this reason, I have deactivated
multiprocess execution for the time being using a flag.

However, it seems possible to provide a
different storage system to dagster, one that hopefully
doesn't rely on the file system. I think if we manage to do that,
we'll have a fast multiprocess execution.

Note that for multiprocess execution to work, dagster needs to be
able to recreate the pipeline in any new process it spawns. This
imposes some restrictions for the implementation of the new
SpineEngineExperimental:

1. Pipeline creation needs to be done by a function
at module top-level (not by a class method).
2. We cannot rely on the state of ExecutableItem
to pass information between backward and forward pipeline, as the
ExecutableItem instances are rebuilt in each process rather
than shared across. Everything needs to be communicated through solid
Outputs and for this reason we have a single pipeline that does
both the backward and forward sweep.
3. The pipeline can't have access to SpineEngineExperimental
attributes, notably the publisher. This complicates event streaming.
The solution is to yield instances of custom AssetMaterialization
subclasses porting the messages. These are captured by
execute_pipeline_iterator (which runs in the SpineEngineExperimental
instance) so it's possible to dispatch them as events that toolbox
can listen.