Get rid of the termination thread
We were spawning daemon threads to terminate other threads,
but that doesn't make sense and feels dangerous. In the
multiprocess executor that we copied, this works better because
the daemon thread is spawned from a child process that's about to
die, so the daemon dies with the process. In our multithread
executor it looks like daemons were living forever.

Anyways, we manage our own interruption scheme
`SpineEngine.stop` that doesn't rely on KeyBoardInterrupt for
instance, so we can get rid of some of multiprocess executor
machinery. The hope is that all this helps with performance by
make our execution framework a bit lighter.