Add support for filtering db resources over pipeline 'edges'
SpineEngineExperimental now has a new keyword argument,
`filter_stacks`. This argument is used to specify filters for db
resources that are applied when the resources are passed to a
certain item. In this way we can support scenario filtering and
fanning in a generic way, that all project items can utilize
indistinctively.