Fix filter_stacks_iterator
I think this needs to be like before, that is, when a resource has
no filters, the filter stack is a list with an empty dict. Perhaps
the name filter stack is not very consistent with some other concept
I don't know, but here the filter stack is actually a dict
mapping filter type (scenario or tool) to a list of values (scenario
or tool names). So an empty dict means no filters for both
scenario and tool.

The issue with making the filter stack an empty list becomes evident
in the next couple of lines, when we call itertools.product
on the filter stacks yielded by the iterator. If any of this
stacks is empty, the product will also be empty. So the empty
stack needs to be the neutral element of the product operator,
that is, and turns out to be a list with empty dict.

Anyways, without this, filter execution doesn't work if one of
the resources has no filter.