Verify that configs are non-empty before applying them
Because of the previous change, we need to handle the empty
dict appearing as one of the filter configurations in the stacks.
I don't know if there's a better way to do this, but I think
it's good to be able to use itertools.product, even if we need
to work around it a little bit.