Handle empty filter stacks
Because of how we handle things in toolbox, it is possible that
one resource has an empty stack. This happens when e.g. the user
defines filters and then empties the db for some reason. In this case,
we don't immediately clean up the filters although we could...
the result is that some filter stacks arrive empty to engine.

It is very simple to handle the issue in engine so...

We also take the opportunity to fix some minor issue with the filter id
and do some very minor refactoring.

Bump version.