Fix excel_reader
Reader's mapping sniffing algorithm create_mappings_from_sheet() was
unconditionally expecting that the sheets were either empty or contained
a cell with 'sheet_type' string. This is in general not the case for
arbitrary Excel files. We now give up with sniffing if the Excel file
does not look like our official exporter database.