Make multithread executor compatible with dagster 0.10.1
Re #18

We simply drop the setp_key argument in calls to engine_event,
as we don't really care too much for it.

We also set 0.10.1 as dagster upper bound so this doesn't happen
again. The problem is we might miss important upgrades from dagster,
so we need to keep a watching eye.