Use sane default value for the success flag
There are cases when items utterly fail to execute, e.g. when they raise
an unhandled exception. When this happens, the success flag is not
updated leaving it to its default value. Previously, this value was
"UNDEFINED" which caused engine to raise another exception crashing the
execution. The success flag is now set to
ItemExecutionFinishState.NEVER_FINISHED by default so engine won't crash
anymore and we have a way to identify crashed item executions.

Re Spine-project/Spine-Toolbox#1428