Make project_dir in test dicts an absolute path and some clean up
Simplified the extraction path and made sure that it is deleted after running the whole test suite, the test class, or individual tests.
Previously these tests created a directory ~/hellow on Linux and an extra directory into C-drive root on Windows

Changed the project_dir in test dicts to absolute paths because it is always an absolute path in practice.