Limit parallel persistent processes
It is now possible to limit the number of persistent processes that run
simultaneously. The useful 'lifetime' of a persistent process is
between the instantiation of PersistentExecutionManager and the end of
PersistentExecutionManagerBase.run_until_complete(). After that the
process may be killed by Engine if the process limit has been reached.

We use semaphores to keep count on running persistent processes. This
commit also replaces the hand-written semaphore implementation for
ReturningProcess by Python's standard Semaphore class.

Re #35, Spine-project/Spine-Toolbox#1586