Fix resource limits with parallel engines
There was a problem with spine_engine._set_resource_limits(): if two or
more engines were started at the same time (as is the case with multiple
DAGs in a project), every SpineEngine instance would recreate the
semaphores guarding the process limits. This replaces the stock
threading.Semaphore objects by our own semaphore-like objects that make
it more convenient to set and change the limits between multiple
executions and with parallel engines.

Re #35