Do not close socket from SpineREPL after flushing
Re Spine-project/SpineOpt.jl#453

It looks like in some windows setups, the connection is immediately
and irremediably reset after calling close(sock), even if the buffer
is not empty. So if the other end is not 'fast enough' so as to
drain the buffer before that call, we just get a ConnectionResetError
in python.