Stop cleaning connection settings in to_dict()
We were dropping nonexistent but disabled filter names in
ResourceConvertingConnection.to_dict(). The behavior is undesirable,
however, as in Toolbox' headless mode we just can't do that because the
connection's resources are not properly set there. The cleaning step has
now been moved to a proper place in spinetoolbox.

Re Spine-project/Spine-Toolbox#1694