Allow option to check in and out dbs when opening resources
Re Spine-project/Spine-Toolbox#1091

We need to do the check in and out for Tool items ourselves, I don't
like relying on the Tool author to do this. This means that Tools will
lock an output DB during their entire execution which may reduce
parallelization but it's fine in my opinion.