Don't run iterating steps that depend on a failed one
If an item within a loop failed on iteration two or later, their
successors within the loop would continue executing after it. This is
because as the item was successful in the first iteration, their
successors would of course be candidates for execution.

And if all of the successors succeeded, the loop condition would be
valuated once again and the workflow eventually continue looping.