Make dagster pipelines based on selected items and their direct predecessors and successors only
Dagster pipelines included every single item in the DAG even in cases when only a subset was selected for execution.
This resulted in execution being slower than it needed to be and some confusion on why it looked like it was executing
also items that were not even selected. This commit fixes the problem by not filling up the Dagster pipelines with items
that should not be touched.

Also, refactored SpineEngine class to hopefully make it more readable and ran black.

Re spine-tools/Spine-Toolbox#1990