Fix scenario and group ID execution using Jupyter Consoles
We used (kernel_name, group_id) tuple for storing consoles. When executing scenarios, this
tuple was always the same, which meant that only a single Jupyter console was ever launched.
This is fixed by using the filter_id for storing the kernel managers. In addition, Jupyter Consoles 
did not support Group ID's, which is also fixed now.

In addition:
- Updated unit test Github action runner to support Jupyter Console testing
- Added tests for  kernel_execution_manager.py module.

Re spine-tools/Spine-Toolbox#2080
PR:  #96