Ensure all required project items are included in execution
Since some time ago Engine was changed such that we do not instantiate
every project item in a DAG but only the ones that are permitted
for execution and their immediate predecessors. Turns out, however,
that this breaks many a workflow as the item and its predecessor
may not know that they're part of filtered execution and don't get
filter_ids which in turn breaks certain project items that need the
id even in exclude_execution().

With the changes in this commit, we now instatiate all upstream
items up to a filtered connection from permitted items. This makes
Engine deal with filter_ids as intended.

Re #99