Fix race condition when acquiring an idle persistent manager
It is possible that multiple threads have collected idle persistent
managers in _PersistentManagerFactory.new_persistent_manager() at
the same time. If there is no manager that can be reused but
there is an idle manager, one of the threads deletes that manager
and the other threads will then Traceback because the idle manager's
key is not anymore in _PersistentManagerFactory.persistent_managers
dict.

We now use a lock to allow a single thread to execute the idle
manager code to fix the race condition.